html, xml: parameter type changed: int -> size_t added language/encoding detection for oeb/epub files
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@2576 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/src/formats/chm/HHCReader.cpp
@@ -96,7 +96,7 @@ bool HHCReader::tagHandler(const HtmlTag &tag) {
 	return true;
 }
 
-bool HHCReader::characterDataHandler(const char*, int, bool) {
+bool HHCReader::characterDataHandler(const char*, size_t, bool) {
 	return true;
 }
