+++ fbreader/src/formats/dummy/DummyBookReader.cpp
@@ -26,7 +26,7 @@ DummyBookReader::DummyBookReader(BookModel &model) : myModelReader(model) {
 }
 
 /*
-void DummyBookReader::characterDataHandler(const char *text, int len) {
+void DummyBookReader::characterDataHandler(const char *text, size_t len) {
 }
 
 void DummyBookReader::startElementHandler(int tag, const char **xmlattributes) {
