+++ fbreader/src/formats/html/HtmlBookReader.h
@@ -50,11 +50,11 @@ protected:
 	void startDocumentHandler();
 	void endDocumentHandler();
 	bool tagHandler(const HtmlTag &tag);
-	bool characterDataHandler(const char *text, int len, bool convert);
+	bool characterDataHandler(const char *text, size_t len, bool convert);
 
 private:
-	void preformattedCharacterDataHandler(const char *text, int len, bool convert);
-	void addConvertedDataToBuffer(const char *text, int len, bool convert);
+	void preformattedCharacterDataHandler(const char *text, size_t len, bool convert);
+	void addConvertedDataToBuffer(const char *text, size_t len, bool convert);
 
 protected:
 	BookReader myBookReader;
