+++ fbreader/src/formats/openreader/ORBookReader.h
@@ -37,7 +37,7 @@ public:
 
 	void startElementHandler(const char *tag, const char **attributes);
 	void endElementHandler(const char *tag);
-	void characterDataHandler(const char *text, int len);
+	void characterDataHandler(const char *text, size_t len);
 
 private:
 	enum ReaderState {
