+++ zlibrary/text/src/hyphenation/ZLTextHyphenationReader.cpp
@@ -20,7 +20,7 @@
 #include "ZLTextHyphenationReader.h"
 #include "ZLTextTeXHyphenator.h"
 
-void ZLTextHyphenationReader::characterDataHandler(const char *text, int len) {
+void ZLTextHyphenationReader::characterDataHandler(const char *text, size_t len) {
 	if (myReadPattern) {
 		myBuffer.append(text, len);
 	}
